manifest: add bases support to create_manifest() (CRAFT-63)
- Move create_manifest() out of utils into its own module.  This will
prevent a potential import loop due to dependencies on util in config.

- Add optional bases_config parameter to use the run-on bases
configuration.

Note that runtime usage still only invokes the "legacy" non-bases
configuration.  Future work will enable the usage of bases and this
just adds test coverage for the future case.
