+++ charmcraft/commands/pack.py
@@ -22,9 +22,9 @@ from argparse import Namespace
 
 from charmcraft.cmdbase import BaseCommand, CommandError
 from charmcraft.commands import build
+from charmcraft.manifest import create_manifest
 from charmcraft.utils import (
     SingleOptionEnsurer,
-    create_manifest,
     load_yaml,
     useful_filepath,
 )
